+++ src/components/cost-value-big-trip-template.js
@@ -1,0 +1,11 @@
+const createCostValueBigTripTemplate = () => {
+  return (
+    `<p class="trip-info__cost">
+      Total: &euro;&nbsp;<span class="trip-info__cost-value">1230</span>
+    </p>`
+  );
+};
+
+export {
+  createCostValueBigTripTemplate
+};
