+++ src/components/list-big-trip-template.js
@@ -1,0 +1,10 @@
+const createListBigTripTemplate = () => {
+  return (
+    `<ul class="trip-days">
+      </ul>`
+  );
+};
+
+export {
+  createListBigTripTemplate
+};
